Proof of concept for webpack based .60 import
The rustwasmtest example works now differently.

First the rust build needs to be complete:

    cargo build

Then the webpack loader needs to be built:

    cd api/sixtyfps-web/webpack-loader
    npm install
    npm run build

Finally the rustwasmtest example should compile:

    cd examples/rustwasmtest
    npm install

to install the deps, then

    npm run start

will do the compilation and start the web-server with live-reload. The contained index.js will import a .60 file.
webpack is configured to use our own loader plugin to compile the .60 file to wasm.

This is still just starting the run-time instead of exposing an API, but
the infrastructure is there to build on that.